+++ Src/GraphicsEngineOpenGL/scene/Model.cpp
@@ -28,7 +28,7 @@ auto Model::get_aabb() -> std::shared_ptr<AABB> { return aabb; }
 
 auto Model::get_materials() const -> std::vector<ObjMaterial> { return materials; }
 
-auto Model::get_texture_count() const -> int { return static_cast<uint32_t>(texture_list.size()); }
+auto Model::get_texture_count() const -> int { return static_cast<int>(texture_list.size()); }
 
 void Model::load_model_in_ram(const std::string &model_path)
 {
@@ -66,16 +66,16 @@ void Model::create_render_context()
 void Model::bind_ressources()
 {
     glBindBufferBase(GL_SHADER_STORAGE_BUFFER, STORAGE_BUFFER_MATERIAL_ID_BINDING, ssbo);
-    for (int i = 0; std::cmp_less(i, texture_list.size()); i++) {
-        texture_list[i]->use_texture(i + MODEL_TEXTURES_SLOT);
+    for (size_t i = 0; std::cmp_less(i, texture_list.size()); i++) {
+        texture_list[i]->use_texture(static_cast<GLuint>(i) + MODEL_TEXTURES_SLOT);
     }
 }
 
 void Model::unbind_resources()
 {
     glBindBuffer(GL_SHADER_STORAGE_BUFFER, 0);
-    for (int i = 0; std::cmp_less(i, texture_list.size()); i++) {
-        texture_list[i]->unbind_texture(i + MODEL_TEXTURES_SLOT);
+    for (size_t i = 0; std::cmp_less(i, texture_list.size()); i++) {
+        texture_list[i]->unbind_texture(static_cast<GLuint>(i) + MODEL_TEXTURES_SLOT);
     }
 }
